Movies Recommnded System
This project provides a user-friendly interface for exploring different movies and testing different classifiers. It allows users to experiment with hyperparameters to see how they affect classification accuracy. The visualization of data in 2D using PCA adds an insightful visual aspect to understand the dataset's distribution. Users can gain a better understanding of the data and the performance of various classifiers on different datasets.